Correccion de las interfaces y servicios en el tipo de datos, ademas creacion de un paquete co.com.gblock.servicio.utilidad, donde se creo una clase RegistroBodega, donde se pone en funcionamiento el servicio de Bodega.

+++ src/java/co/com/gblock/ImplementacionServicios/ArqueoServicio.java
@@ -15,6 +15,6 @@ import javax.jws.WebService;
  */
 @Stateless
 @WebService(endpointInterface="co.com.gblock.InterfaceServicios.IArqueoServicio")
-public class ArqueoServicio extends GenericaServicio<Arqueo, Integer> implements IArqueoServicio{
+public class ArqueoServicio extends GenericaServicio<Arqueo, Long> implements IArqueoServicio{
     
 }
